+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/EvaluationExplorer/components/index.ts
@@ -1,0 +1,5 @@
+export { DatasetSection } from './DatasetSection';
+export { ModelsSection } from './ModelsSection';
+export { ScorersSection } from './ScorersSection';
+export { EditToolbar } from './EditToolbar';
+export { EvaluationDataGrid } from './EvaluationDataGrid';
