+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/EvaluationExplorer/index.ts
@@ -1,0 +1,2 @@
+export { EvaluationExplorerPage } from './EvaluationExplorerPage';
+export type { EvaluationExplorerPageProps } from './types';
